+++ src/main/java/io/github/torand/fastersql/dialect/Capability.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Tore Eide Andersen
+ * Copyright (c) 2024-2025 Tore Eide Andersen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -47,5 +47,10 @@ public enum Capability {
     /**
      * Supports "nulls First" and "nulls last" for NULL ordering
      */
-    NULL_ORDERING
+    NULL_ORDERING,
+
+    /**
+     * Supports "SELECT ... FOR UPDATE" clauses
+     */
+    SELECT_FOR_UPDATE
 }