+++ src/main/java/io/github/torand/fastersql/dialect/OracleDialect.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Tore Eide Andersen
+ * Copyright (c) 2024-2025 Tore Eide Andersen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,6 +22,7 @@ import java.util.Optional;
 import static io.github.torand.fastersql.dialect.Capability.CONCAT_OPERATOR;
 import static io.github.torand.fastersql.dialect.Capability.LIMIT_OFFSET;
 import static io.github.torand.fastersql.dialect.Capability.NULL_ORDERING;
+import static io.github.torand.fastersql.dialect.Capability.SELECT_FOR_UPDATE;
 import static io.github.torand.fastersql.util.lang.StringHelper.generate;
 
 /**
@@ -33,7 +34,7 @@ public class OracleDialect implements Dialect {
     private final EnumSet<Capability> supportedCaps;
 
     public OracleDialect() {
-        this(EnumSet.of(LIMIT_OFFSET, CONCAT_OPERATOR, NULL_ORDERING));
+        this(EnumSet.of(LIMIT_OFFSET, CONCAT_OPERATOR, NULL_ORDERING, SELECT_FOR_UPDATE));
     }
 
     private OracleDialect(EnumSet<Capability> capabilities) {
@@ -110,8 +111,23 @@ public class OracleDialect implements Dialect {
     }
 
     @Override
+    public String formatRoundFunction(String operand) {
+        return "round(" + operand + ")";
+    }
+
+    @Override
     public String formatModuloFunction(String divisor, String dividend) {
         return "mod(" + divisor + ", " + dividend + ")";
+    }
+
+    @Override
+    public String formatCurrentDateFunction() {
+        return "current_date";
+    }
+
+    @Override
+    public Optional<String> getConcatOperator() {
+        return Optional.of("||");
     }
 
     @Override