+++ src/main/java/io/github/torand/fastersql/function/system/CurrentDate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Tore Eide Andersen
+ * Copyright (c) 2024-2025 Tore Eide Andersen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ public class CurrentDate implements SystemFunction {
 
     @Override
     public String sql(Context context) {
-        return "current_date";
+        return context.getDialect().formatCurrentDateFunction();
     }
 
     // Projection