+++ src/test/java/io/github/torand/fastersql/statement/oracle/OracleInsertBatchStatementTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Tore Eide Andersen
+ * Copyright (c) 2024-2025 Tore Eide Andersen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@ package io.github.torand.fastersql.statement.oracle;
 import io.github.torand.fastersql.domainmodel.Product;
 import io.github.torand.fastersql.domainmodel.ProductCategory;
 import io.github.torand.fastersql.statement.PreparableStatement;
-import io.github.torand.fastersql.util.RowValueMatchers;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collection;
@@ -72,7 +71,7 @@ public class OracleInsertBatchStatementTest extends OracleTest {
                 "PR_DESCRIPTION", is("TBD"),
                 "PR_CATEGORY", is("FURNITURE"),
                 "PR_PRICE", isBigDecimal(1234.56),
-                "PR_STOCK_COUNT", RowValueMatchers.isBigDecimal(46)
+                "PR_STOCK_COUNT", isBigDecimal(46)
             )
             .assertRow(2,
                 "PR_ID", is(id2.toString()),
@@ -80,7 +79,7 @@ public class OracleInsertBatchStatementTest extends OracleTest {
                 "PR_DESCRIPTION", isNull(),
                 "PR_CATEGORY", is("APPLIANCE"),
                 "PR_PRICE", isBigDecimal(4567.89),
-                "PR_STOCK_COUNT", RowValueMatchers.isBigDecimal(34)
+                "PR_STOCK_COUNT", isBigDecimal(34)
             )
             .assertRow(3,
                 "PR_ID", is(id3.toString()),
@@ -88,7 +87,7 @@ public class OracleInsertBatchStatementTest extends OracleTest {
                 "PR_DESCRIPTION", is("Power and portability"),
                 "PR_CATEGORY", is("ELECTRONICS"),
                 "PR_PRICE", isBigDecimal(9012.34),
-                "PR_STOCK_COUNT", RowValueMatchers.isBigDecimal(9)
+                "PR_STOCK_COUNT", isBigDecimal(9)
             )
             .verify(
                 select(PRODUCT.ID, PRODUCT.NAME, PRODUCT.DESCRIPTION, PRODUCT.CATEGORY, PRODUCT.PRICE, PRODUCT.STOCK_COUNT)